requirements.txt
Flask
twilio
python-dotenv
langchain-groq
langchain-core
langchain-community
groq
gunicorn
requests
